Add support for custom `http.Client` and improve documentation  (#10)
* Support Custom Client for NewClient

NewClient should accept an *http.Client, allowing users to provide things like a desired timeout. If the same defaultClient is desired, that can be passed as the parameter (i.e. in api_test.go).
Updated function doc comment to reflect this change.

* Update Go Doc Comments

Adding additional doc comments to functions and structs in the api package.

---------

Co-authored-by: Jonathan Bowe <jbowe@webstaurantstore.com>